refactor(common): Removed circular dependencies
