+++ packages/common/pipes/default-value.pipe.ts
@@ -1,4 +1,5 @@
-import { ArgumentMetadata, Injectable, PipeTransform } from '../index';
+import { Injectable } from '../decorators/core/injectable.decorator';
+import { ArgumentMetadata, PipeTransform } from '../interfaces/features/pipe-transform.interface';
 import { isNil } from '../utils/shared.utils';
 
 /**
