+++ packages/common/pipes/parse-uuid.pipe.ts
@@ -1,6 +1,7 @@
-import { Optional } from '../decorators';
-import { ArgumentMetadata, HttpStatus, Injectable } from '../index';
-import { PipeTransform } from '../interfaces/features/pipe-transform.interface';
+import { Injectable } from '../decorators/core/injectable.decorator';
+import { Optional } from '../decorators/core/optional.decorator';
+import { HttpStatus } from '../enums/http-status.enum';
+import { ArgumentMetadata, PipeTransform } from '../interfaces/features/pipe-transform.interface';
 import {
   ErrorHttpStatusCode,
   HttpErrorByCode,
